requirements.txt
tqdm==4.60.0
h5py==3.1.0
matplotlib==3.4.1
numba==0.53.1
numpy==1.19.2
scipy==1.6.2
PyYAML==6.0
tensorboard==2.4.1
torchac==0.9.3
plyfile==0.7.4
pygame==2.1.2